+++ app/models/project.rb
@@ -14,6 +14,7 @@ class Project < ApplicationRecord
   validates :advisory_board_date, presence: true, on: :create
   validates :advisory_board_date, past_date: true
   validates :establishment_sharepoint_link, presence: true, url: {hostnames: SHAREPOINT_URLS}, on: :create
+  validates :trust_sharepoint_link, presence: true, url: {hostnames: SHAREPOINT_URLS}, on: :create
 
   validate :first_day_of_month
   validate :target_completion_date_is_in_the_future, on: :create